requirements.txt
numpy
PyAutoGUI
PyGetWindow
Pillow
pynput